Fixed typos and moved _handle_rerun function to orchestration init
